feat : fixture 생성 후 경로 수정
