+++ packages/blog/src/apis/useGetPosts/useGetPosts.test.tsx
@@ -1,27 +1,12 @@
-import { describe, test, expect, vi } from 'vitest';
-import { PostData } from '@recoil/postsAtom';
+import { describe, test, expect } from 'vitest';
 import useGetPosts, { adapter } from './useGetPosts';
 import { renderTestHook } from '@utils/testUtil';
 import { graphqlMockError } from '@mock/server';
 import { waitFor } from '@testing-library/react';
+import { MOCK_GET_POSTS_RESPONSE } from '@fixtures/posts';
 
 const renderUseGetPosts = () => {
   return renderTestHook({ hook: useGetPosts });
-};
-
-interface PostsResponseData {
-  posts: PostData[];
-}
-
-const MOCK_DATA: PostsResponseData = {
-  posts: [
-    {
-      id: 1,
-      postKey: 'msw-1',
-      title: 'title1',
-      tags: ['tag1', 'tag2'],
-    },
-  ],
 };
 
 describe('useGetPosts 테스트', () => {
@@ -31,7 +16,9 @@ describe('useGetPosts 테스트', () => {
 
       expect(result.current.isLoading).toBe(true);
 
-      await waitFor(() => expect(result.current.postsData).toEqual(adapter(MOCK_DATA)));
+      await waitFor(() =>
+        expect(result.current.postsData).toEqual(adapter(MOCK_GET_POSTS_RESPONSE)),
+      );
 
       expect(result.current.isLoading).toBe(false);
     });
