+++ packages/blog/src/mock/handlers.ts
@@ -1,32 +1,15 @@
+import { MOCK_GET_POST_WITH_KEY_RESPONSE, MOCK_GET_POSTS_RESPONSE } from '@fixtures/posts';
 import { graphql, HttpResponse } from 'msw';
-
-// todo : mock data 정리
 
 const graphqlHandlers = [
   graphql.query('GetPosts', () => {
     return HttpResponse.json({
-      data: {
-        posts: [
-          {
-            id: 1,
-            postKey: 'msw-1',
-            title: 'title1',
-            tags: ['tag1', 'tag2'],
-          },
-        ],
-      },
+      data: MOCK_GET_POSTS_RESPONSE,
     });
   }),
   graphql.query('GetPostWithKey', () => {
     return HttpResponse.json({
-      data: {
-        post: {
-          id: 1,
-          postKey: 'msw-1',
-          title: 'title1',
-          tags: ['tag1', 'tag2'],
-        },
-      },
+      data: MOCK_GET_POST_WITH_KEY_RESPONSE,
     });
   }),
 ];
