+++ packages/blog/src/recoil/postsAtom/atom.ts
@@ -1,11 +1,5 @@
+import { PostData } from '@fixtures/posts';
 import { atom } from 'recoil';
-
-export interface PostData {
-  id: number;
-  postKey: string;
-  title: string;
-  tags: string[];
-}
 
 const postsAtom = atom<PostData[]>({
   key: 'postAtom',
