+++ packages/blog/src/recoil/postsAtom/index.ts
@@ -1,4 +1,1 @@
-import { PostData } from './atom';
-
 export { default } from './atom';
-export type { PostData };
